+++ R/waver.R
@@ -1,5 +1,4 @@
-#' @import sp
-#' @import rgdal
+#' @import sf
 #' @importFrom methods as is
 #' @importFrom stats setNames
 NULL